2.2.22-rc3
o 3ware IDE raid small update (Adam Radford)
o Fix incorrect comments (Solar Designer)
o Sanity check in isdn (Solar Designer)
o Type fixes for usb (Solar Designer)
o Vmalloc corner case fix (Dave Miller)

+++ drivers/block/loop.c
@@ -334,8 +334,9 @@ static int create_missing_block(struct loop_device *lo, int block, int blksize)
 	set_fs(old_fs);
 
 	if (retval < 0) {
-		printk(KERN_WARNING "loop: cannot create block - FS write failed: code %d\n", 
-									retval);
+		printk(KERN_WARNING
+		    "loop: cannot create block - FS write failed: code %d\n",
+		    (int)retval);
 		return FALSE;
 	} else {
 		return TRUE;
